* Chaged the filters to use floats after running some tests and finding out that the use of doubles made no difference to the results.  * Fixed a few bugs with the sampler and refactored after doing some profiling.

+++ src/GafferImage/Filter.cpp
@@ -46,19 +46,19 @@
 namespace GafferImage
 {
 
-Filter::Filter( double radius, double scale )
+Filter::Filter( float radius, float scale )
 	: m_radius( radius )
 {
 	setScale( scale );
 }
 
-void Filter::setScale( double scale )
+void Filter::setScale( float scale )
 {
-	m_scale = std::max( scale, 1. ); 
+	m_scale = std::max( scale, 1.f ); 
 	m_scaledRadius = m_radius * m_scale;
 }
 
-FilterPtr Filter::create( const std::string &name, double scale )
+FilterPtr Filter::create( const std::string &name, float scale )
 {
 	// Check to see whether the requested Filter is registered and if not, throw an exception.
 	std::vector<std::string>::iterator it( filterList().begin() );
